# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "QMS SaaS Pro API",
    "description": "ISO 13485 / 21 CFR Part 11 compliant QMS REST API.\n\nMulti-tenant SaaS — all data is isolated by `organization_id` via RLS.\nAuthentication uses Supabase Auth with httpOnly cookies.",
    "version": "1.0.0",
    "contact": {
      "name": "QMS SaaS Pro",
      "url": "https://github.com/mohammedboubekeurmm320-sudo/smq-iso-13485-pro"
    }
  },
  "servers": [
    {
      "url": "https://localhost:3000",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sb-<project-ref>-auth-token"
      }
    },
...
  },
  "paths": {
...
    "/api/auth/verify-signature": {
      "post": {
        "summary": "Verify electronic signature",
        "description": "Verifies the user's password before generating an electronic signature. Required for 21 CFR Part 11 §11.100 compliance.",
        "security": [{ "cookieAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["password", "recordId"],
                "properties": {
                  "password": { "type": "string" },
                  "recordId": { "type": "string" },
                  "signatureType": { "type": "string", "enum": ["approval", "rejection", "review", "verification"] }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Signature verified",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "signatureHash": { "type": "string", "description": "HMAC-SHA256 hex (64 chars)" },
                    "signedAt": { "type": "string", "format": "date-time" },
                    "signatureType": { "type": "string" },
                    "signedById": { "type": "string", "format": "uuid" },
                    "signerName": { "type": "string" },
                    "signerRole": { "type": "string" }
                  }
                }
              }
            }
          },
          "401": { "description": "Invalid password" }
        }
      }
    },
...
  }
}